+++ analysis/src/main/java/org/hps/analysis/tuple/SVTTimingTupleDriver.java
@@ -21,7 +21,7 @@ import org.hps.recon.tracking.ShapeFitParameters;
 import org.hps.recon.tracking.TrackData;
 import org.hps.recon.tracking.TrackType;
 import org.hps.recon.tracking.TrackerHitUtils;
-import org.hps.recon.utils.TrackClusterMatcher;
+import org.hps.recon.utils.TrackClusterMatcherNSigma;
 import org.hps.record.triggerbank.SSPData;
 import org.lcsim.detector.converter.compact.subdetector.SvtStereoLayer;
 import org.lcsim.detector.tracker.silicon.HpsSiSensor;
@@ -57,7 +57,7 @@ public class SVTTimingTupleDriver extends MCTupleMaker {
     private String fittedTrackerHitCollectionName = "SVTFittedRawTrackerHits";
     private String trackerHitCollectionName = "StripClusterer_SiTrackerHitStrip1D";
     String[] fpQuantNames = {"nEle_per_Event", "nPos_per_Event", "nPhoton_per_Event", "nUnAssociatedTracks_per_Event", "avg_delX_at_ECal", "avg_delY_at_ECal", "avg_E_Over_P", "avg_mom_beam_elec", "sig_mom_beam_elec"};
-    private TrackClusterMatcher matcher = new TrackClusterMatcher();
+    private TrackClusterMatcherNSigma matcher = new TrackClusterMatcherNSigma();
     //private Map<SiSensor, Map<Integer, Hep3Vector>> stripPositions = new HashMap<SiSensor, Map<Integer, Hep3Vector>>(); 
     private List<HpsSiSensor> sensors = null;
     private Map<Integer, List<SvtStereoLayer>> topStereoLayers = new HashMap<Integer, List<SvtStereoLayer>>();
